+++ src/test/java/blackjack/domain/participant/DealerTest.java
@@ -3,8 +3,8 @@ package blackjack.domain.participant;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import blackjack.domain.CardTest;
-import blackjack.domain.Deck;
+import blackjack.domain.card.CardTest;
+import blackjack.domain.card.Deck;
 import blackjack.domain.card.Card;
 import blackjack.domain.card.Shape;
 import blackjack.domain.card.Value;
