+++ src/test/java/blackjack/domain/participant/PlayersTest.java
@@ -4,7 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-import blackjack.domain.Deck;
+import blackjack.domain.card.Deck;
 import java.util.Collections;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
